# docs/openapi.yaml
openapi: 3.1.0
info:
  title: vet402 API
  version: 1.0.0
  description: |
    Independent verification for the x402 agent-payment economy. Scores payee
    wallets and ERC-8004 agents on Base, returns a 0-100 score with an
    ALLOW/WARN/BLOCK recommendation, and publishes what happened after those
    verdicts.

    **Disclaimer:** Scores are informational only and do not constitute a guarantee,
    credit assessment, or investment advice. Every response that carries a score or
    a published rate repeats this string in the payload, so the caveat travels with
    the number.

    **Rate limits.** Authenticated calls: a per-account monthly quota (Free 1,000
    / Pro 50,000 / Scale 500,000), shared across keys, reset at the UTC month
    boundary. Each `/score` is 1 unit; a `/scores/batch` of N agents is N units.
    Scored responses carry `X-RateLimit-Limit`, `X-RateLimit-Used`, and
    `X-RateLimit-Remaining`. There is no per-second burst throttle on
    authenticated calls — you may spend the quota as fast as you like; a `429`
    with `Retry-After` means the monthly quota is exhausted. Separate IP-based
    abuse throttles (e.g. 60 auth-failures/min/IP) exist only to blunt
    key-guessing and scraping and are not hit by valid authenticated traffic.

    Key-less calls are throttled per IP instead. Every response on those paths
    carries `RateLimit-Limit` — the ceiling for that path, true for every caller.
...

...
    response.

  contact:
    name: vet402 API Support
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /api/v1/agents/{agentId}/score:
    get:
      operationId: getAgentScore
      tags: [Scores]
      summary: Get trust score by ERC-8004 agent ID
      deprecated: true
      description: |
        Returns trust score for a registered ERC-8004 agent.
        Pass `wallet` query param to verify the agent's registered wallet matches.

        Deprecated since 2026-09-02: use
        `GET /api/v1/resources/{resourceId}/decision`. The score's value is
        unchanged; since 2026-09-29 the response says so itself
        (`deprecated`, `deprecation`, and the `Deprecation` / `Link` headers).
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/VerifyWallet"
      responses:
        "200":
          description: Trust score result
          headers:
            Deprecation:
              $ref: "#/components/headers/ScoreDeprecation"
            Link:
              $ref: "#/components/headers/ScoreDeprecationLink"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ScoreDeprecationNotice"
                  - $ref: "#/components/schemas/TrustScoreResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/payees/{address}/score:
    get:
      operationId: getPayeeScore
      tags: [Payees]
      summary: Get trust score for a payee (payment-receiving wallet)
      deprecated: true
      description: |
        Deprecated since 2026-09-02: use
        `GET /api/v1/resources/{resourceId}/decision?role=payer` for the
        endpoint you are about to pay (`GET /api/v1/resolve?q={url}` finds its
        `resourceId`). This score rates the wallet as a whole, not that
        endpoint, and can say ALLOW where the decision says WARN or BLOCK.
        Its value is unchanged; since 2026-09-29 the response says so itself
        (`deprecated`, `deprecation`, and the `Deprecation` / `Link` headers).

        Buyer-side counterpart to `/api/v1/agents/{agentId}/score` and
        `/api/v1/wallets/{address}/score`,
        which screen the caller of an x402 route. This endpoint screens the *recipient*:
        "should my agent pay this wallet?" before an x402 payment is sent.

        Combines the wallet's own settlement receiving history (payments attested via
        `POST /api/v1/payments/x402` where it was the payee), wallet-age/tx-count health, an
        exit-scam-shaped outflow check (received funds then drained them), and any prior
        `verdict_outcomes` labels naming this wallet.

        Never 404s for an unfamiliar wallet — a wallet with little or no receiving
        history still returns `200` with `dataDepth: "thin"`, so integrators can decide
        how much to trust a thin-data score rather than being blocked outright.
      parameters:
        - $ref: "#/components/parameters/WalletAddress"
      responses:
        "200":
          description: Payee score result
          headers:
            Deprecation:
              $ref: "#/components/headers/ScoreDeprecation"
            Link:
              $ref: "#/components/headers/ScoreDeprecationLink"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ScoreDeprecationNotice"
                  - $ref: "#/components/schemas/PayeeScoreResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/observatory/export.csv:
    get:
      operationId: exportObservatoryLedgerCsv
      tags: [Observatory]
      summary: Purchase-ledger CSV export (audit-trail serialization)
      description: |
        The same paid-attempt series the endpoint pages publish, as CSV —
        a portable audit trail (Phase 4.2 v0). No new claims are minted here:
        rows are the ledger facts (attempted_at, resource_key, network,
        status, amounts, tx_hash, delivery figures). Excludes vet402-side
        non-events (budget_denied / halted / request_error / in_flight). `days`
        saturates to 1..366; output truncates at 50,000 rows and says so in
        the `x-vet402-truncated` header. IP-throttled to 6/min.

        Columns, in order (columns are only ever appended; existing names,
        order and meaning do not change): `attempted_at`, `resource_key`,
        `network`, `status`, `amount_units`, `spent_units`, `tx_hash`,
        `http_status_paid`, `latency_ms`, `l2_schema`, `held_reason`
        (since 2026-09-17), `request_body`, `request_body_sha256`,
        `settlement_source` (since 2026-09-20), `request_query`,
        `request_query_sha256` (since 2026-09-21), `confirmed_units`
        (since 2026-09-29).

        - `amount_units` — the price the seller's 402 asked for, in the
          asset's base units (USDC: 6 decimals).
        - `spent_units` — the amount vet402 signed for on this attempt: what
          it put at stake, not an amount shown to have moved. A
          `settle_failed` row with no `tx_hash` still carries the price here.
          Read `confirmed_units` for money that moved. The meaning of this
          column has not changed; the note is new.
        - `confirmed_units` — the amount vet402 re-read on-chain as
          transferred: equal to `spent_units` on `settled` rows and `0` on
          every other row, including `settle_claimed` rows that are still
          awaiting the re-read (they take the amount once they turn
          `settled`). `0` means no transfer vet402 has confirmed, not a
          transfer shown not to have happened. The response also carries
          both definitions in the `x-vet402-column-notes` header, because a
          CSV has nowhere to put a comment.

        - `held_reason` — why the row is held rather than counted against the
          seller: `settled_4xx`, `unsettled_4xx`, `payer_unfunded`; blank when
          the row is counted.
        - `request_body` — what the paid request carried: `declared` (the body
          the seller's own 402 declares in `extensions.bazaar.info.input.body`,
          sent unedited; since 2026-09-17), `empty` (`{}`), `none` (a request
          formed with no body, whether or not it reached the seller; recorded
          since 2026-09-20). Blank means the row
          holds no record: rows before 2026-09-17, bodiless requests before
          2026-09-20, and rows that ended before a paid request went out.
          Blank is not `empty`. The row does not record the HTTP method, so
          older rows are not back-filled.
        - `request_body_sha256` — on `declared` rows since 2026-09-20, the
          lowercase hex SHA-256 of the exact bytes sent (the declared body
          passed through `JSON.stringify`). The body itself is not
          republished; it is the seller's text and the seller's 402 serves it
          to anyone. Blank otherwise.
        - `settlement_source` — who named `tx_hash`: `seller_claim` (the paid
          response carried a settlement receipt naming it) or `vet402_index`
          (no receipt came back; vet402's own settlements index found the
          transfer by payer, payee, exact amount and time window, and the row
          then went through the same on-chain re-read). Blank when the row has
          no `tx_hash`. The settled count of `vet402_index` rows is
          `l1.settledLateLinked` on `/api/v1/observatory/state`.
        - `request_query` — which URL the paid request went to: `declared`
          (the query arguments the seller's own 402 declares in
          `extensions.bazaar.info.input.queryParams`, appended to the listed
          URL as declared), `empty` (the seller declared none), `refused` (a
          declaration was there and vet402's own rules did not use it — a
          non-scalar value, an empty name, over the caps, declared names that
          collide once folded, a host or path that would move, or a collision
          with names the listed URL already carries that left nothing to
          append). A declared name the listed URL already carries is dropped
          on its own: the listing's query stands and the declaration's other
          pairs are still appended, so that row is `declared` and the hash
          below says what went out. The label has been recorded on the row
          since 2026-09-20 and published as this column since 2026-09-21, on
          the networks named in vet402's per-network allow-list; the paid
          request's URL and the row are unchanged on every other network.
          Blank means the row holds no record: a row on a network the
          allow-list does not name, a Tempo (MPP) row, a row that ended before
          a paid request went out, and rows before 2026-09-20. The column
          alone does not separate those four from one another, and blank is
          not `empty`. Older rows are not back-filled. Within the window the
          response covers, the rows whose `request_query` is not blank are the
          networks this has been in effect on, and the `declared` ones are
          where pairs were actually appended; a network can be on the
          allow-list and carry nothing but `empty` and `refused`.
        - `request_query_sha256` — on `declared` rows, the lowercase hex
          SHA-256 of the appended pairs written as one form-urlencoded string:
          the appended pairs alone (neither the query the listed URL already
          carried nor a declared name dropped for colliding with one), in the
          declaration's own key order (integer-like names first, ascending, by
          the JavaScript rule), `application/x-www-form-urlencoded` so a space
          is `+`, `&` between pairs and `=` between name and value, and no
          leading `?` or `&`. The query string itself is not republished: the
          names and values are the seller's own, and the 402 that answered
          vet402's unpaid request is where they were read from. The hash covers the appended pairs alone, not the listed URL or
          which endpoint was bought, so it matches two rows on the same
          endpoint, or a row against a fresh 402; it is not a way of hiding a
          short declared query. Blank otherwise.
        - `confirmed_units` — see above: `spent_units` on `settled` rows,
          `0` otherwise.

        `scripts/declared-body-before-after.mjs` recomputes, from this CSV
        alone, the paid 2xx rate of the same endpoints before and after the
        declared body.
      security: []
      parameters:
        - in: query
          name: days
          required: false
          schema: { type: integer, minimum: 1, maximum: 366, default: 90 }
      responses:
        "200":
          description: CSV, oldest attempt first
          content:
            text/csv: {}
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: Rate limited
        "503":
          description: Observatory store unavailable
...
components:
  headers:
    ScoreDeprecation:
      description: RFC 9745 deprecation date of the 0–100 score (2026-09-02T00:00:00Z).
      schema: { type: string, example: "@1788307200" }
    ScoreDeprecationLink:
      description: Where the deprecation is explained (`rel="deprecation"`).
      schema: { type: string, example: '<https://vet402.com/docs/api#endpoints>; rel="deprecation"; type="text/html"' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key issued from the developer dashboard
  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      description: ERC-8004 agent token ID (uint256 as decimal string)
      schema:
        type: string
        example: "42"
    WalletAddress:
      name: address
      in: path
      required: true
      description: EVM wallet address (checksummed or lowercase)
      schema:
        type: string
        pattern: "^0x[a-fA-F0-9]{40}$"
        example: "0x1234567890123456789012345678901234567890"
    VerifyWallet:
      name: wallet
      in: query
      required: false
      description: Optional wallet to verify against agent's registered wallet
      schema:
        type: string
        pattern: "^0x[a-fA-F0-9]{40}$"
  schemas:
    ScoreDeprecationNotice:
      type: object
      description: |
        Carried by every 0–100 score response since 2026-09-29. The score
        itself is unchanged; this says it is deprecated and what to call instead.
      required: [deprecated, deprecation]
      properties:
        deprecated:
          type: boolean
          enum: [true]
        deprecation:
          type: object
          required: [since, use_instead, find_resource_id, message]
          properties:
            since: { type: string, format: date, example: "2026-09-02" }
            use_instead: { type: string }
            find_resource_id: { type: string }
            message: { type: string }
    Recommendation:
      type: string
      enum: [ALLOW, WARN, BLOCK]
...
    TrustScoreResult:
      type: object
      required:
        - agentId
        - trustScore
        - recommendation
        - signals
        - scoredAt
        - cacheExpiresAt
        - disclaimer
      properties:
        agentId:
          type: string
          description: ERC-8004 agent ID; "0" if wallet-only lookup
        wallet:
          type: string
          nullable: true
        trustScore:
          type: integer
          minimum: 0
          maximum: 100
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        signals:
          $ref: "#/components/schemas/TrustSignals"
        breakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        reasons:
          type: array
          items: { type: string }
          description: >-
            Stable machine-readable codes behind the verdict, derived from the
            same signals (for integrators' compliance logs). Cannot disagree
            with the recommendation.
        scoredAt:
          type: string
          format: date-time
        cacheExpiresAt:
          type: string
          format: date-time
        disclaimer:
          type: string
        manualOverride:
          type: boolean
          description: True when customer WL/BL changed the recommendation
        blockReason:
          type: string
          description: Present when recommendation is BLOCK (e.g. wallet_mismatch, operator_policy)
        dataCoverage:
          type: object
          description: Indexer and settlement data freshness
          properties:
            ownerIndexer:
              type: object
              properties:
                status:
                  type: string
                  enum: [synced, partial, unavailable]
                blocksBehind:
                  type: integer
                  nullable: true
                lastBlock:
                  type: string
                  nullable: true
                  description: Last block the owner indexer has processed (decimal string); null when unknown
                indexedAgentRows:
                  type: integer
                  description: Agent rows the owner indexer currently holds
                staleRisk: { type: boolean }
            settlement:
              type: object
              properties:
                paymentRows: { type: integer }
                distinctWallets: { type: integer }
                recentPayments30d: { type: integer }
                walletHasHistory: { type: boolean }
...
    ScoreBreakdown:
      type: object
      description: >-
        N-21 explainability. Decomposition of the chain score into its four
        weighted components plus the sybil adjustment — the exact arithmetic of
        the weighting, so it cannot disagree with trustScore. Omitted on
        hard-blocked verdicts where no weighting ran. The manual whitelist/
        blacklist layer is intentionally excluded (see manualOverride).
      required: [components, weightedSubtotal, sybilPenalty, prePolicyScore]
      properties:
        components:
          type: object
          required: [identity, reputation, wallet, x402]
          properties:
            identity: { $ref: "#/components/schemas/ScoreComponent" }
            reputation: { $ref: "#/components/schemas/ScoreComponent" }
            wallet: { $ref: "#/components/schemas/ScoreComponent" }
            x402: { $ref: "#/components/schemas/ScoreComponent" }
        weightedSubtotal:
          type: number
          description: Weighted average of the four components before the sybil penalty
        sybilPenalty:
          type: number
          description: Points removed by sybil/availability flags (<= 0); see signals.sybil.flags
        prePolicyScore:
          type: integer
          minimum: 0
          maximum: 100
          description: weightedSubtotal + sybilPenalty, clamped; equals trustScore unless a manual list moved it
...
    PayeeScoreResult:
      type: object
      required:
        - payee
        - score
        - recommendation
        - dataDepth
        - degraded
        - signalsUnavailable
        - signals
        - scoredAt
        - cacheExpiresAt
        - disclaimer
      properties:
        payee:
          type: string
          description: Lowercased wallet address that was scored
        score:
          type: integer
          minimum: 0
          maximum: 100
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        dataDepth:
          $ref: "#/components/schemas/DataDepth"
        degraded:
          type: boolean
          description: >-
            **Read this before you act on `recommendation`.** True when at
            least one input could not be read at all, so this body is a
            fail-closed refusal, not a measurement. `dataDepth` answers "how
            much history does this wallet have?"; this answers "did we manage
            to look?".

            A client MUST NOT treat a `degraded: true` body as ALLOW, whatever
            `recommendation` and `score` say. The engine already forces
            BLOCK here and refuses to cache the verdict; @vet402/sdk
            (SpendGuard, `payee_score_degraded`) and @vet402/middleware
            (`score_degraded`) both deny on this field alone, and so does the
            Python client in the repository (`packages/python-sdk`, not
            published to PyPI). The `@vouchscore/*` packages are the former,
            deprecated names of the same npm packages. Always present.
        signalsUnavailable:
          type: array
          items: { type: string }
          description: >-
            **Read this before you act on `recommendation`.** Every input that
            could not be read on this request, named — currently one or more of
            `wallet_metrics`, `native_drain`, `usdc_drain`, `outcome_history`.
            Empty means the whole assessment was measured. Treat the list as
            open: new input names may be added without a breaking change.

            Non-empty with `degraded: false` is the partial-measurement case:
            the score IS backed by real measurements, just not all of them, and
            it has been capped below ALLOW for exactly that reason. A client
            that must not act on an incomplete view MUST NOT treat a body with
            a non-empty `signalsUnavailable` as ALLOW — SpendGuard denies it as
            `payee_partial_measurement` and the middleware as
            `partial_measurement`. Always present (empty array when clean).
        signals:
          $ref: "#/components/schemas/PayeeSignals"
        scoredAt:
          type: string
          format: date-time
        cacheExpiresAt:
          type: string
          format: date-time
        disclaimer:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: |
        Rate limit exceeded. On authenticated paths this means the monthly quota
        is spent (`rate_limit_exceeded`); on key-less paths it is the per-IP
        window (`rate_limited`).
      headers:
        Retry-After:
          description: Seconds until the window resets.
          schema: { type: integer }
        RateLimit-Limit:
          description: Key-less paths — ceiling for the IP window. Sent on every response.
          schema: { type: integer }
        RateLimit-Remaining:
          description: >-
            Key-less paths — requests left in your current IP window. Per-caller,
            so it is sent here and on other uncached responses, but not on a
            CDN-cached 200.
          schema: { type: integer }
        RateLimit-Reset:
          description: >-
            Key-less paths — unix epoch seconds when your window resets. Same
            per-caller rule as RateLimit-Remaining.
          schema: { type: integer }
        X-RateLimit-Limit:
          description: Authenticated paths — monthly plan quota.
          schema: { type: integer }
        X-RateLimit-Used:
          description: Authenticated paths — quota consumed this month.
          schema: { type: integer }
        X-RateLimit-Remaining:
          description: Authenticated paths — quota left this month.
          schema: { type: integer }
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: rate_limit_exceeded
              retryAfter:
                type: integer